borra si es admin

+++ api/controllers/ComentariosApiController.php
@@ -37,52 +37,59 @@ class ComentariosApiController extends Api {
         //     return $this->json_response($response, 200);
         // }
         
-        function InsertTarea($param = null){
+        // function InsertTarea($param = null){
             
-            $objetoJson = $this->getJSONData();
-            $r = $this->model->InsertarTarea($objetoJson->Titulo, $objetoJson->Descripcion, $objetoJson->Completada);
+            //     $objetoJson = $this->getJSONData();
+            //     $r = $this->model->InsertarTarea($objetoJson->Titulo, $objetoJson->Descripcion, $objetoJson->Completada);
             
-            return $this->json_response($r, 200);
-        }
-        
-        function postComentario($param) {
-            if (isset($param)) {
-                $id_accion = $param[0];
-                // $body = json_decode($this->raw_data);
-                $body = $this->getData();
-                $usuario = $body->usuario;
-                $id_usuario = $this->usuariosModel->getID('usuario', $usuario);
-                $titulo = $body->titulo;
-                $cuerpo = $body->cuerpo;
-                $puntaje = $body->puntaje;
-                
-                $comentarioNuevo = $this->comentariosModel->postComentario($titulo, $cuerpo, $puntaje, $id_accion, $id_usuario);
-                if (!$comentarioNuevo) {
-                    return $this->json_response("comentario vacío", 300);
+            //     return $this->json_response($r, 200);
+            // }
+            
+            function postComentario($param) {
+                if (isset($param)) {
+                    $id_accion = $param[0];
+                    // $body = json_decode($this->raw_data);
+                    $body = $this->getData();
+                    $usuario = $body->usuario;
+                    $id_usuario = $this->usuariosModel->getID('usuario', $usuario);
+                    $titulo = $body->titulo;
+                    $cuerpo = $body->cuerpo;
+                    $puntaje = $body->puntaje;
+                    
+                    $comentarioNuevo = $this->comentariosModel->postComentario($titulo, $cuerpo, $puntaje, $id_accion, $id_usuario);
+                    if (!$comentarioNuevo) {
+                        return $this->json_response("comentario vacío", 300);
+                    }
+                    else {
+                        return $this->json_response($comentarioNuevo, 200);
+                    }
                 }
                 else {
-                    return $this->json_response($comentarioNuevo, 200);
+                    return $this->json_response("Server Error", 500);
                 }
             }
-            else {
-                return $this->json_response("Server Error", 500);
-            }
-        }
-        
-        function deleteComentario($param) {
-            if (isset($param)) {
-                $id_comentario = $param[0];
-                $result = $this->comentariosModel->deleteComentario($id_comentario);
-                if (!$result) {
+            
+            function deleteComentario($param) {
+                session_start();
+                if (isset($_SESSION['user'])) {
+                    $isAdmin = $this->usuariosModel->isAdmin($_SESSION['user']);
+                }
+                else {
+                    $isAdmin = false;
+                }
+                if (isset($param) && $isAdmin) {
+                    $id_comentario = $param[0];
+                    $result = $this->comentariosModel->deleteComentario($id_comentario);
+                    if (!$result) {
+                        return $this->json_response("no especifica accion", 300);
+                    }
+                    else {
+                        return $this->json_response($result, 200);
+                    }
+                    
+                }
+                else {
                     return $this->json_response("no especifica accion", 300);
                 }
-                else {
-                    return $this->json_response($result, 200);
-                }
-                
-            }
-            else {
-                return $this->json_response("no especifica accion", 300);
             }
         }
-    }
